Divide classes into pacakges
