+++ src/main/java/duke/Event.java
@@ -1,18 +1,21 @@
+package duke;
+
 import java.time.LocalDate;
-public class Deadline extends Task {
+public class Event extends Task {
     protected String type;
     protected LocalDate time;
 
-    Deadline(String name, LocalDate time) {
+    public Event(String name, LocalDate time) {
         super(name);
         this.time = time;
-        this.type = "D";
+        this.type = "E";
     }
 
     @Override
     public String toString() {
         String status = this.isDone ? "X" : " ";
-        return "[" + this.type + "][" + status + "]" + this.name + " (by: "
+        return "[" + this.type + "][" + status + "]" + this.name + " (at: "
                 + DateTimeParser.formatDate(this.time) + ")";
     }
+
 }
