+++ src/main/java/duke/Task.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.io.File;
 import java.util.ArrayList;
