+++ src/main/java/duke/exceptions/InvalidTypeException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class InvalidTypeException extends Exception {
     public InvalidTypeException(String message) {
         super(message);
