+++ src/main/java/duke/exceptions/MissingDeadlineDateException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class MissingDeadlineDateException extends Exception {
     public MissingDeadlineDateException(String message) {
         super(message);
