+++ src/main/java/duke/exceptions/MissingEventDateException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class MissingEventDateException extends Exception{
     public MissingEventDateException(String message) {
         super(message);
