+++ src/main/java/duke/exceptions/MissingNameException.java
@@ -1,3 +1,5 @@
+package duke.exceptions;
+
 public class MissingNameException extends Exception {
     public MissingNameException(String message) {
             super(message);
